bug fixes, code refactor
